packages/support-bot/requirements.txt
# Support Bot Dependencies
# LangChain/LangGraph core
langgraph>=0.2.50
langchain>=0.3.27
langchain-core>=0.3.75

# LLM Providers  
langchain-openai>=0.3.29
langchain-google-genai>=2.0.0

# Vector Store for RAG
langchain-chroma>=0.1.2
chromadb>=0.5.15

requests

# Data processing
pydantic>=2.0.0
pydantic-settings>=2.0.0

# AWS Lambda utilities (matching existing project patterns)
aws-lambda-powertools[tracer,logger,metrics]>=2.25.0

# HTTP and async support
httpx>=0.27.0
aiohttp>=3.9.0

# Environment management
python-dotenv>=1.0.0

# Optional: for local testing and development
fastapi>=0.104.0
uvicorn>=0.24.0

